Stop double auditing
At the very end of passwd if we are changing the password, it calls
pam_chauthtok(). This function already emits an audit record with
pass/fail indication. As it stands, pam sends an audit event saying
the password was changed and then passwd sends an audit event
saying the password was changed. This patch removes the final audit
events and places a comment that pam is responsible for reporting the
outcome.

Signed-off-by: sgrubb <sgrubb@redhat.com>